Refactor scripts names
