+++ scripts/src/getDemoMeta.ts
